tambahan kedua

+++ database/migrations/2025_01_16_164333_create_employees_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama');
             $table->enum('jeniskelamin',['cowo','cewe']);
-            $table->integer('notelepon');
+            $table->bigInteger('notelepon');
             $table->timestamps();
         });
     }
